backend/requirements.txt
fastapi
starlette_session
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
PyJWT
python-multipart
sendgrid
authlib
httpx
geoalchemy2
alembic
supabase
twilio